+++ api/db.js
@@ -1,8 +1,6 @@
 const express = require ("express");
 const mongoose = require ("mongoose");
 const dotenv = require ("dotenv");
-const morgan = require ("morgan"); 
-const app = express();
 
 dotenv.config();
 
@@ -15,3 +13,5 @@ mongoose.connect(process.env.MONGO_URL,
         console.log("********CONEXION CORRECTA****")
     }
 });
+
+module.exports = mongoose;
