+++ api/index.js
@@ -1,33 +1,7 @@
-const express = require ("express");
-
-const mongoose = require ("mongoose");
+const app = require('./app')
+const db = require('./db')
 const dotenv = require ("dotenv");
-const morgan = require ("morgan"); 
-
-const app = express();
 
 dotenv.config();
 
-require('./db');
-
-const { Router } = require('express');
-const hostRoute = require ('./src/routes/host')
-const bookingRoute = require ('./src/routes/booking')
-const guestRoute = require ('./src/routes/guest')
-
-  //middleware
-  app.use(express.json());
-  app.use(morgan("common"));
-
-  app.use("/api/host", hostRoute);
-  app.use("/api/booking", bookingRoute);
-  app.use("/api/guest", guestRoute);
- 
-
-
-
-
-
-
-
-app.listen(3000, ()=> {console.log ("listening on port 3000")})
+app.listen(process.env.PORT, ()=> {console.log ("listening on port 3001")})
