+++ api/package.json
@@ -14,6 +14,9 @@
     "express": "^4.18.1",
     "mongoose": "^6.5.2",
     "morgan": "^1.10.0",
-    "nodemon": "^2.0.19"
+    "nodemon": "^2.0.19",
+    "body-parser": "^1.20.0",
+    "cookie-parser": "^1.4.6",
+    "cors": "^2.8.5"
   }
 }
